move some structs around
