+++ azure-kusto-data/src/prelude.rs
@@ -13,4 +13,5 @@
 
 pub use crate::client::{KustoClient, KustoClientOptions};
 pub use crate::connection_string::ConnectionStringBuilder;
+pub use crate::operations::mgmt::KustoResponseDataSetV1;
 pub use crate::operations::query::{KustoResponseDataSetV2, ResultTable};
